+++ .vscode/launch.json
@@ -8,8 +8,8 @@
       "args": [],
       "stopAtEntry": false,
       "externalConsole": false,
-      "cwd": "/home/dakotath/Documents/wii_projects/wii-examples/threading/basic_threading/source",
-      "program": "/home/dakotath/Documents/wii_projects/wii-examples/threading/basic_threading/source/build/Debug/outDebug",
+      "cwd": "/home/dakotath/Documents/wii_projects/wii-examples/threading/source",
+      "program": "/home/dakotath/Documents/wii_projects/wii-examples/threading/source/build/Debug/outDebug",
       "MIMode": "gdb",
       "miDebuggerPath": "gdb",
       "setupCommands": [
